% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{add_group_brackets}
\alias{add_group_brackets}
\title{Add group brackets}
\usage{
add_group_brackets(p, group_bracket_df, bracket_x, bracket_arm)
}
\arguments{
\item{p}{ggplot object, the current plot}

\item{group_bracket_df}{a data frame with y coordinates for each of the grouping brackets}

\item{bracket_x}{}
}
\value{
p, a ggplot object, with the added red brackets for each of the grouped taxa
}
\description{
Add group brackets
}
