% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{add_selected_vr_rects}
\alias{add_selected_vr_rects}
\title{Add selected v-region rectangles}
\usage{
add_selected_vr_rects(p, selected_vr_rects)
}
\arguments{
\item{p}{ggplot object, the current plot}

\item{selected_vr_rects}{a tibble with x and y coordinates in order to draw each of the selected v-region rectangles}
}
\value{
p, a ggplot object, with the added golden rectangles drawn at the coordinates provided in selected_vr_rects
}
\description{
Add selected v-region rectangles
}
