% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{add_tile_ids}
\alias{add_tile_ids}
\title{Add tile ids}
\usage{
add_tile_ids(p, data, label_col, x_col, y_col, enabled, size)
}
\arguments{
\item{p}{ggplot object, the current plot}

\item{data}{a data frame with the tiles and their ids}

\item{x_col}{-- x coordinates for each of the 9 v-region columns}

\item{y_col}{lsit of y coordinates for each of the tile rows}

\item{enabled}{boolean for whether tile ids are enabled by the user}

\item{size}{int -- font size}
}
\value{
p, a ggplot object, with the added green checkmarks for each of the grouped taxa
}
\description{
Add tile ids
}
