% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{add_unique_taxa_checks}
\alias{add_unique_taxa_checks}
\title{Add unique taxa checkmarks}
\usage{
add_unique_taxa_checks(p, unique_taxa_df, x, y_col)
}
\arguments{
\item{p}{ggplot object, the current plot}

\item{unique_taxa_df}{a data frame with all unique taxa}

\item{x}{int -- the x coordinate of the checkmarks on the plot}

\item{y_col}{lsit of y coordinates for each of the checkmarks}
}
\value{
p, a ggplot object, with the added green checkmarks for each of the grouped taxa
}
\description{
Add unique taxa checkmarks
}
