% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{build_selected_vr_rects}
\alias{build_selected_vr_rects}
\title{Build selected v-region rectangles}
\usage{
build_selected_vr_rects(selected_vr, vr_levels_all, ymin, ymax)
}
\arguments{
\item{selected_vr}{list of selected v-regions}

\item{vr_levels_all}{list of all v-regions}

\item{ymin}{int for the top of the rectangles}

\item{ymax}{int for the bottom of the rectangles}
}
\value{
a tibble with x and y coordinates in order to draw each of the selected v-region rectangles
}
\description{
Build selected v-region rectangles
}
