% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{build_species_layout}
\alias{build_species_layout}
\title{Build the species_layout dataframe for the visualization}
\usage{
build_species_layout(RADqtiles, groups_info, gap = 2)
}
\arguments{
\item{RADqtiles}{...}

\item{groups_info}{...}

\item{gap}{A changeable int that determines the vertical gap between species in a column}
}
\value{
a list containing species_layout, copy_layout, RADqtiles, y_breaks, group_bracket_df, and unique_taxa_df dfs
}
\description{
Build the species_layout dataframe for the visualization
}
