% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_loading.R
\name{get_loading_messages}
\alias{get_loading_messages}
\title{Get loading screen messages}
\usage{
get_loading_messages()
}
\value{
A character vector of loading messages.
}
\description{
Get loading screen messages
}
