% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_organism_selection_list.R
\name{load_organism_selection_list}
\alias{load_organism_selection_list}
\title{Load list used in RADexplorer dropdown}
\usage{
load_organism_selection_list(accessions_table)
}
\arguments{
\item{accessions_table}{a table with columns "organism" and "genus"}
}
\description{
Load list used in RADexplorer dropdown
}
