% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_loading.R
\name{loading_screen_ui}
\alias{loading_screen_ui}
\title{Build the RADexplorer loading screen UI}
\usage{
loading_screen_ui()
}
\value{
A bslib page_fillable UI object containing a centered loading
spinner and a randomly selected loading message.
}
\description{
Build the RADexplorer loading screen UI
}
