% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{make_plotly_layout}
\alias{make_plotly_layout}
\title{Build plotly layout}
\usage{
make_plotly_layout(p_msa, plot_height)
}
\arguments{
\item{p_msa}{a ggplot object of the visualization}

\item{plot_height}{Int - height of the plot}
}
\value{
A plotly object of the final visualization
}
\description{
Build plotly layout
}
