% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{make_species_axis_labels}
\alias{make_species_axis_labels}
\title{Build species y-axis labels}
\usage{
make_species_axis_labels(species, n_copies, searched_taxa = character(0))
}
\arguments{
\item{species}{List of user-selected genus species names}

\item{n_copies}{List of number of copies}
}
\value{
a list containing species labels for the y axis
}
\description{
Build species y-axis labels
}
