% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_menu.R
\name{menu_screen_ui}
\alias{menu_screen_ui}
\title{Build the RADexplorer menu screen UI}
\usage{
menu_screen_ui()
}
\value{
A bslib page_fillable UI object containing the taxa selection menu.
}
\description{
Build the RADexplorer menu screen UI
}
