% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_radx.R
\name{radx_instructions}
\alias{radx_instructions}
\title{Build the RADx instructions accordion panel}
\usage{
radx_instructions()
}
\value{
A bslib accordion UI object containing usage instructions.
}
\description{
Build the RADx instructions accordion panel
}
