% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_radx.R
\name{radx_loading_plot}
\alias{radx_loading_plot}
\title{Build a plotly output with a loading overlay}
\usage{
radx_loading_plot(id, height = "570px")
}
\arguments{
\item{id}{Character string; the output ID for the plotly plot.}

\item{height}{Character string; the CSS height of the plot, default "570px".}
}
\value{
A div containing the plot output and loading spinner overlay.
}
\description{
Build a plotly output with a loading overlay
}
