% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_radx.R
\name{radx_screen_ui}
\alias{radx_screen_ui}
\title{Build the main RADexplorer screen UI}
\usage{
radx_screen_ui()
}
\value{
A bslib page_sidebar UI object containing the sidebar controls and main plot area.
}
\description{
Build the main RADexplorer screen UI
}
