% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen_radx.R
\name{radx_styles}
\alias{radx_styles}
\title{Build shared CSS styles for the RADx screen}
\usage{
radx_styles()
}
\value{
A shiny tags$style object containing CSS for the RADx UI components.
}
\description{
Build shared CSS styles for the RADx screen
}
