% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_helpers.R
\name{standardize_plot_inputs}
\alias{standardize_plot_inputs}
\title{Standardize plot inputs for RADexplorer visualizations}
\usage{
standardize_plot_inputs(RADq, unique, groups, varRegions)
}
\arguments{
\item{RADq}{Data frame of RADq results.}

\item{unique}{Data frame of summarized unique IDs.}

\item{groups}{Data frame of taxa grouping information.}

\item{varRegions}{List of the user-selected variable regions}
}
\value{
a list containing the standardized unique, RADqtiles, groups_info, and copy_counts
}
\description{
Standardize plot inputs for RADexplorer visualizations
}
